Fixed app errors

+++ packages/backend/endpoints/notes/index.js
@@ -1,5 +1,6 @@
 
 import pool from '../../databases/postgres/index.js';
+import getUserIdFromCookie from '../../utils/getUserIdFromCookie/index.js';
 
 /**
  * Handles notes endpoints.
@@ -14,16 +15,22 @@ class NotesEndpoints {
     try {
       const client = await pool.connect();
 
-      /** @type { import('../shared/interfaces/index.js').AddNoteProps } */
+      /** @type { import('../../../shared/interfaces/index.js').CreateNoteRequestProps } */
       const body = req.body;
 
-      console.log(body)
+      const { userId, message } = await getUserIdFromCookie(req);
+
+      console.log("User ID:", userId);
+
+      if (userId === 0) {
+        return res.status(401).json({ success: false, message: message });
+      }
 
       const now = Date.now();
-      const newNoteQuery = 'INSERT INTO public.note (user_id, title, content, created_at, updated_at) VALUES ($1, $2, $3, $4, $5) RETURNING id;';
-      const newNoteValues = [body.userID, body.title, body.content, now, now];
+      const query = 'INSERT INTO public.note (user_id, title, content, created_at, updated_at) VALUES ($1, $2, $3, $4, $5) RETURNING id;';
+      const values = [userId, body.title, body.content, now, now];
 
-      const { rows } = await client.query(newNoteQuery, newNoteValues);
+      const { rows } = await client.query(query, values);
 
       await client.release(true);
 
@@ -45,11 +52,14 @@ class NotesEndpoints {
     try {
       const client = await pool.connect();
 
-      /** @type {{ userID: number }} */
-      const body = req.body;
+      const { userId, message } = await getUserIdFromCookie(req);
+
+      if (userId === 0) {
+        return res.status(401).json({ success: false, message: message });
+      }
 
       const fetchNotesQuery = 'SELECT * FROM public.note WHERE user_id = $1;';
-      const fetchNotesValues = [body.userID];
+      const fetchNotesValues = [userId];
 
       const { rows } = await client.query(fetchNotesQuery, fetchNotesValues);
 
@@ -73,20 +83,21 @@ class NotesEndpoints {
     try {
       const client = await pool.connect();
 
-      /** @type {import('../shared/interfaces/index.js').ReadNoteProps} */
-      const body = req.body;
+      const { noteId } = req.params;
+      const { userId, message } = await getUserIdFromCookie(req);
 
-      console.log("HEADERS", req.headers);
-      console.log("COOKIES", req.cookies);
+      if (userId === 0) {
+        return res.status(401).json({ success: false, message: message });
+      }
 
       const fetchNoteQuery = 'SELECT * FROM public.note WHERE id = $1 AND user_id = $2 LIMIT 1;';
-      const fetchNoteValues = [body.id, body.userID];
+      const fetchNoteValues = [noteId, userId];
 
       const { rows } = await client.query(fetchNoteQuery, fetchNoteValues);
 
       await client.release(true);
 
-      return res.status(200).json({ success: true, message: 'Fetched note successfully', note: JSON.stringify(rows[0]) });
+      return res.status(200).json({ success: true, message: 'Fetched notes successfully', note: JSON.stringify(rows[0]) });
 
     } catch (err) {
       console.error(err);
@@ -104,14 +115,21 @@ class NotesEndpoints {
     try {
       const client = await pool.connect();
 
-      /** @type {import('../shared/interfaces/index.js').EditNoteProps} */
+      /** @type {import('../../../shared/interfaces/index.js').UpdateNoteRequestProps} */
       const body = req.body;
 
-      const now = Date.now();
-      const updateNoteQuery = 'UPDATE public.note SET title = $1, content = $2, updated_at = $3 WHERE id = $4;';
-      const updateNoteValues = [body.title, body.content, now, body.id];
+      const { noteId } = req.params;
+      const { userId, message } = await getUserIdFromCookie(req);
 
-      await client.query(updateNoteQuery, updateNoteValues);
+      if (userId === 0) {
+        return res.status(401).json({ success: false, message: message });
+      }
+
+      const now = Date.now();
+      const query = 'UPDATE public.note SET title = $1, content = $2, updated_at = $3 WHERE id = $4 AND user_id = $5;';
+      const values = [body.title, body.content, now, noteId, userId];
+
+      await client.query(query, values);
 
       await client.release(true);
 
@@ -133,13 +151,17 @@ class NotesEndpoints {
     try {
       const client = await pool.connect();
 
-      /** @type { import('../shared/interfaces/index.js').RemoveNoteProps } */
-      const body = req.body;
+      const { noteId } = req.params;
+      const { userId, message } = await getUserIdFromCookie(req);
 
-      const deleteNoteQuery = 'DELETE FROM public.note WHERE id = $1 AND user_id = $2';
-      const deleteNoteValues = [body.id, body.userID];
+      if (userId === 0) {
+        return res.status(401).json({ success: false, message: message });
+      }
 
-      await client.query(deleteNoteQuery, deleteNoteValues);
+      const query = 'DELETE FROM public.note WHERE id = $1 AND user_id = $2';
+      const values = [noteId, userId];
+
+      await client.query(query, values);
 
       await client.release(true);
